split Content.tsx
