+++ src/components/Stepper/components/Content/Success/types.ts
@@ -1,0 +1,5 @@
+import type { Step } from '../../../../../features/data/reducers/wallet/stepper-types.ts';
+
+export type SuccessContentProps = {
+  step: Step;
+};
